all: switch to a registration system
In the style of database/sql.

I am not actually a fan of registration systems.
The old code was written more the way I like to see Go code written:
explicit about what is used. The downside is unlike say, database/sql
where creating new DB objects is rare, it's very common to create a
treesitter parser. So this required importing two packages all over
the place, and it is hard to think of a good name for the language
packages. Simply "golang" is very confusing, but "treesittergo" or
"gotreesitter" is very long. So do the dull thing and make a global map.